refactored utils to src folders
